requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
fake-useragent==0.1.11
fastapi==0.82.0
h11==0.13.0
idna==3.3
lxml==4.9.1
outcome==1.2.0
pydantic==1.10.2
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.4.3
selenium-stealth==1.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.19.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
wsproto==1.2.0
